test(schedule): split routes.test.ts along source seams
The 1,954-line godfile becomes three seam-aligned files plus one shared
harness (SPEC ceiling: no test file over ~1,200 lines):

- routes-jobs.test.ts (923) — lib/routes/jobs.ts + the job-service verbs
- exec-tools.test.ts (747) — lib/exec-tools.ts (bakin_exec_schedule_*)
- activation-and-search.test.ts (408) — search indexing, GET /search,
  registration, ensureBakinJob provisioning
- helpers/schedule-harness.ts — pure fixture builders + runtime-cron mock
  state factory; mock.module wiring stays per-file (FW2 B7 precedent)

Pure relocation: 98 pass / 6 skip / 332 expect() calls before and after.
Each split file carries the dual content-dir mock with a full-shape
getBakinPaths (the real execution ledger resolves its path via .db).
